requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.0
openai>=1.50.0
anthropic>=0.39.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
chromadb>=0.5.0  # optional: for vector-based RAG. Falls back to keyword retrieval without it.
